+++ examples/website/arc/app.js
@@ -2,7 +2,8 @@
 import React, {Component} from 'react';
 import {render} from 'react-dom';
 import MapGL from 'react-map-gl';
-import DeckGLOverlay from './deckgl-overlay.js';
+import DeckGL, {MapView, GeoJsonLayer, ArcLayer} from 'deck.gl';
+import {scaleQuantile} from 'd3-scale';
 
 // Set your mapbox token here
 const MAPBOX_TOKEN = process.env.MapboxAccessToken; // eslint-disable-line
@@ -11,32 +12,100 @@ const MAPBOX_TOKEN = process.env.MapboxAccessToken; // eslint-disable-line
 const DATA_URL =
   'https://raw.githubusercontent.com/uber-common/deck.gl-data/master/examples/arc/counties.json'; // eslint-disable-line
 
-class Root extends Component {
+const inFlowColors = [
+  [255, 255, 204],
+  [199, 233, 180],
+  [127, 205, 187],
+  [65, 182, 196],
+  [29, 145, 192],
+  [34, 94, 168],
+  [12, 44, 132]
+];
+
+const outFlowColors = [
+  [255, 255, 178],
+  [254, 217, 118],
+  [254, 178, 76],
+  [253, 141, 60],
+  [252, 78, 42],
+  [227, 26, 28],
+  [177, 0, 38]
+];
+
+const INITIAL_VIEW_STATE = {
+  longitude: -100,
+  latitude: 40.7,
+  zoom: 3,
+  maxZoom: 15,
+  pitch: 30,
+  bearing: 30
+};
+
+class App extends Component {
   constructor(props) {
     super(props);
     this.state = {
-      viewport: {
-        ...DeckGLOverlay.defaultViewport,
-        width: 500,
-        height: 500
-      },
-      data: null,
+      counties: null,
+      arcs: null,
       selectedCounty: null
     };
 
-    fetch(DATA_URL)
-      .then(response => response.json())
-      .then(({features}) => {
-        this.setState({
-          data: features,
-          selectedCounty: features.find(f => f.properties.name === 'Los Angeles, CA')
+    if (!window.demoLauncherActive) {
+      fetch(DATA_URL)
+        .then(response => response.json())
+        .then(({features}) => {
+          this.setState({
+            counties: features,
+            selectedCounty: features.find(f => f.properties.name === 'Los Angeles, CA')
+          });
+          this._recalculateArcs(this.state.counties, this.state.selectedCounty);
+
         });
-      });
+    } else {
+      this._recalculateArcs(this.props.data, this.props.selectedFeature);
+    }
   }
 
   componentDidMount() {
     window.addEventListener('resize', this._resize.bind(this));
     this._resize();
+  }
+
+  componentWillReceiveProps(nextProps) {
+    const arcsChanged =
+      nextProps.data !== this.props.data ||
+      nextProps.selectedFeature !== this.props.selectedFeature;
+    if (arcsChanged) {
+      this._recalculateArcs(nextProps.data, nextProps.selectedFeature);
+    }
+  }
+
+  _recalculateArcs(data, selectedFeature) {
+    if (!data || !selectedFeature) {
+      return;
+    }
+
+    const {flows, centroid} = selectedFeature.properties;
+
+    const arcs = Object.keys(flows).map(toId => {
+      const f = data[toId];
+      return {
+        source: centroid,
+        target: f.properties.centroid,
+        value: flows[toId]
+      };
+    });
+
+    const scale = scaleQuantile()
+      .domain(arcs.map(a => Math.abs(a.value)))
+      .range(inFlowColors.map((c, i) => i));
+
+    arcs.forEach(a => {
+      a.gain = Math.sign(a.value);
+      a.quantile = scale(Math.abs(a.value));
+    });
+
+    this.setState({arcs});
   }
 
   _resize() {
@@ -52,7 +121,9 @@ class Root extends Component {
 
   _onClick(info) {
     // Clicked a county
-    this.setState({selectedCounty: info.object});
+    const selectedCounty = info.object;
+    this.setState({selectedCounty});
+    this._recalculateArcs(this.props.data, selectedCounty);
   }
 
   _onViewportChange(viewport) {
@@ -62,27 +133,65 @@ class Root extends Component {
   }
 
   render() {
-    const {viewport, data, selectedCounty} = this.state;
+    const {
+      strokeWidth = 2,
+      onHover = this._onHover.bind(this),
+      onClick = this._onClick.bind(this),
+
+      onViewStateChange = (({viewState}) => this.setState({viewState})),
+      viewState = this.state.viewState,
+
+      mapboxApiAccessToken = MAPBOX_TOKEN,
+      mapStyle = "mapbox://styles/mapbox/light-v9"
+    } = this.props;
+
+    const layers = [
+      new GeoJsonLayer({
+        id: 'geojson',
+        data: this.state.counties,
+        stroked: false,
+        filled: true,
+        getFillColor: () => [0, 0, 0, 0],
+        onHover,
+        onClick,
+        pickable: Boolean(onHover || onClick)
+      }),
+      new ArcLayer({
+        id: 'arc',
+        data: this.state.arcs,
+        getSourcePosition: d => d.source,
+        getTargetPosition: d => d.target,
+        getSourceColor: d => (d.gain > 0 ? inFlowColors : outFlowColors)[d.quantile],
+        getTargetColor: d => (d.gain > 0 ? outFlowColors : inFlowColors)[d.quantile],
+        strokeWidth
+      })
+    ];
 
     return (
       <MapGL
-        {...viewport}
-        onViewportChange={this._onViewportChange.bind(this)}
-        mapboxApiAccessToken={MAPBOX_TOKEN}
+        {...viewState}
+        reuseMap
+        onViewportChange={viewport => onViewStateChange({viewState: viewport})}
+        mapboxApiAccessToken={mapboxApiAccessToken}
+        mapStyle={mapStyle}
+        preventStyleDiffing={true}
       >
-        <DeckGLOverlay
-          viewport={viewport}
-          data={data}
-          selectedFeature={selectedCounty}
-          onHover={this._onHover.bind(this)}
-          onClick={this._onClick.bind(this)}
-          strokeWidth={2}
-        />
+
+        <DeckGL
+          layers={layers}
+          views={new MapView({id: 'map'})}
+          viewState={viewState}
+          />;
+
       </MapGL>
     );
   }
 }
 
+// NOTE: EXPORTS FOR DECK.GL WEBSITE DEMO LAUNCHER - CAN BE REMOVED IN APPS
+export {App, INITIAL_VIEW_STATE};
+export {inFlowColors, outFlowColors};
+
 if (!window.demoLauncherActive) {
-  render(<Root />, document.body.appendChild(document.createElement('div')));
+  render(<App />, document.body.appendChild(document.createElement('div')));
 }
