+++ examples/website/brushing/app.js
@@ -2,11 +2,16 @@
 import React, {Component} from 'react';
 import {render} from 'react-dom';
 import MapGL from 'react-map-gl';
-import DeckGLOverlay from './deckgl-overlay.js';
+import DeckGL, {MapView} from 'deck.gl';
+import ArcBrushingLayer from './arc-brushing-layer';
+import ScatterplotBrushingLayer from './scatterplot-brushing-layer';
+import {scaleLinear} from 'd3-scale';
+
 
 // Set your mapbox token here
 const MAPBOX_TOKEN = process.env.MapboxAccessToken; // eslint-disable-line
-const tooltipStyle = {
+
+const TOOLTIP_STYLE = {
   position: 'absolute',
   padding: '4px',
   background: 'rgba(0, 0, 0, 0.8)',
@@ -21,27 +26,127 @@ const tooltipStyle = {
 const DATA_URL =
   'https://raw.githubusercontent.com/uber-common/deck.gl-data/master/examples/arc/counties.json'; // eslint-disable-line
 
-class Root extends Component {
+export const inFlowColors = [[35, 181, 184]];
+export const outFlowColors = [[166, 3, 3]];
+
+// migrate out
+const SOURCE_COLOR = [166, 3, 3];
+// migrate in
+const TARGET_COLOR = [35, 181, 184];
+
+const INITIAL_VIEW_STATE = {
+  longitude: -100,
+  latitude: 40.7,
+  zoom: 3,
+  maxZoom: 15,
+  pitch: 0,
+  bearing: 0
+};
+
+class App extends Component {
   constructor(props) {
     super(props);
     this.state = {
-      viewport: {
-        ...DeckGLOverlay.defaultViewport,
-        width: 500,
-        height: 500
-      },
+      arcs: [],
+      targets: [],
+      sources: [],
       data: null,
       mousePosition: [0, 0]
     };
 
-    fetch(DATA_URL)
-      .then(response => response.json())
-      .then(({features}) => this.setState({data: features}));
+    if (!window.demoLauncherActive) {
+      fetch(DATA_URL)
+        .then(response => response.json())
+        .then(({features}) => this.setState({data: features}));
+    }
   }
 
+  /* eslint-disable react/no-did-mount-set-state */
   componentDidMount() {
     window.addEventListener('resize', this._resize.bind(this));
     this._resize();
+
+    this.setState({
+      ...this._getLayerData(this.props)
+    });
+  }
+  /* eslint-enable react/no-did-mount-set-state */
+
+  UNSAFE_componentWillReceiveProps(nextProps) {
+    if (nextProps.data !== this.props.data) {
+      this.setState({
+        ...this._getLayerData(nextProps)
+      });
+    }
+  }
+
+  _getLayerData({data}) {
+    if (!data) {
+      return null;
+    }
+    const arcs = [];
+    const targets = [];
+    const sources = [];
+    const pairs = {};
+
+    data.forEach((county, i) => {
+      const {flows, centroid: targetCentroid} = county.properties;
+      const value = {gain: 0, loss: 0};
+
+      Object.keys(flows).forEach(toId => {
+        value[flows[toId] > 0 ? 'gain' : 'loss'] += flows[toId];
+
+        // if number too small, ignore it
+        if (Math.abs(flows[toId]) < 50) {
+          return;
+        }
+        const pairKey = [i, Number(toId)].sort((a, b) => a - b).join('-');
+        const sourceCentroid = data[toId].properties.centroid;
+        const gain = Math.sign(flows[toId]);
+
+        // add point at arc source
+        sources.push({
+          position: sourceCentroid,
+          target: targetCentroid,
+          name: data[toId].properties.name,
+          radius: 3,
+          gain: -gain
+        });
+
+        // eliminate duplicates arcs
+        if (pairs[pairKey]) {
+          return;
+        }
+
+        pairs[pairKey] = true;
+
+        arcs.push({
+          target: gain > 0 ? targetCentroid : sourceCentroid,
+          source: gain > 0 ? sourceCentroid : targetCentroid,
+          value: flows[toId]
+        });
+      });
+
+      // add point at arc target
+      targets.push({
+        ...value,
+        position: [targetCentroid[0], targetCentroid[1], 10],
+        net: value.gain + value.loss,
+        name: county.properties.name
+      });
+    });
+
+    // sort targets by radius large -> small
+    targets.sort((a, b) => Math.abs(b.net) - Math.abs(a.net));
+    const sizeScale = scaleLinear()
+      .domain([0, Math.abs(targets[0].net)])
+      .range([36, 400]);
+
+    targets.forEach(pt => {
+      pt.radius = Math.sqrt(sizeScale(Math.abs(pt.net)));
+    });
+
+    return {arcs, targets, sources};
   }
 
   _resize() {
@@ -83,19 +188,96 @@ class Root extends Component {
     }
 
     return (
-      <div style={{...tooltipStyle, left: x, top: y}}>
+      <div style={{...TOOLTIP_STYLE, left: x, top: y}}>
         <div>{hoveredObject.name}</div>
         <div>{`Net gain: ${hoveredObject.net}`}</div>
         <div>{`i: ${hoveredObject.i}`}</div>
       </div>
     );
   }
-  render() {
-    const {viewport, data, mousePosition, mouseEntered} = this.state;
 
-    if (!data) {
+  render() {
+    const {
+      enableBrushing = true,
+      brushRadius = 100000,
+      strokeWidth = 2,
+      opacity = 0.7,
+
+      mouseEntered = this.state.mouseEntered,
+      mousePosition = this.sate.mousePosition,
+      onHover = this._onHover.bind(this),
+
+      onViewStateChange = (({viewState}) => this.setState({viewState})),
+      viewState = this.state.viewState,
+
+      mapboxApiAccessToken = MAPBOX_TOKEN,
+      mapStyle = "mapbox://styles/mapbox/dark-v9"
+    } = this.props;
+    const {arcs, targets, sources} = this.state;
+
+    // mouseEntered is undefined when mouse is in the component while it first loads
+    // enableBrushing if mouseEntered is not defined
+    const isMouseover = mouseEntered !== false;
+    const startBrushing = Boolean(isMouseover && enableBrushing);
+
+    if (!arcs || !targets) {
       return null;
     }
+
+    const layers = [
+      new ScatterplotBrushingLayer({
+        id: 'sources',
+        data: sources,
+        brushRadius,
+        brushTarget: true,
+        mousePosition,
+        opacity: 1,
+        enableBrushing: startBrushing,
+        pickable: false,
+        // only show source points when brushing
+        radiusScale: startBrushing ? 3000 : 0,
+        getColor: d => (d.gain > 0 ? TARGET_COLOR : SOURCE_COLOR),
+        getTargetPosition: d => [d.position[0], d.position[1], 0]
+      }),
+      new ScatterplotBrushingLayer({
+        id: 'targets-ring',
+        data: targets,
+        brushRadius,
+        mousePosition,
+        strokeWidth: 2,
+        outline: true,
+        opacity: 1,
+        enableBrushing: startBrushing,
+        // only show rings when brushing
+        radiusScale: startBrushing ? 4000 : 0,
+        getColor: d => (d.net > 0 ? TARGET_COLOR : SOURCE_COLOR)
+      }),
+      new ScatterplotBrushingLayer({
+        id: 'targets',
+        data: targets,
+        brushRadius,
+        mousePosition,
+        opacity: 1,
+        enableBrushing: startBrushing,
+        pickable: true,
+        radiusScale: 3000,
+        onHover: this.props.onHover,
+        getColor: d => (d.net > 0 ? TARGET_COLOR : SOURCE_COLOR)
+      }),
+      new ArcBrushingLayer({
+        id: 'arc',
+        data: arcs,
+        strokeWidth,
+        opacity,
+        brushRadius,
+        enableBrushing: startBrushing,
+        mousePosition,
+        getSourcePosition: d => d.source,
+        getTargetPosition: d => d.target,
+        getSourceColor: d => SOURCE_COLOR,
+        getTargetColor: d => TARGET_COLOR
+      })
+    ];
 
     return (
       <div
@@ -104,28 +286,32 @@ class Root extends Component {
         onMouseLeave={this._onMouseLeave.bind(this)}
       >
         {this._renderTooltip()}
+
         <MapGL
-          {...viewport}
-          onViewportChange={this._onViewportChange.bind(this)}
-          mapboxApiAccessToken={MAPBOX_TOKEN}
+          {...viewState}
+          reuseMap
+          onViewportChange={viewport => onViewStateChange({viewState: viewport})}
+          mapboxApiAccessToken={mapboxApiAccessToken}
+          mapStyle={mapStyle}
+          preventStyleDiffing={true}
         >
-          <DeckGLOverlay
-            viewport={viewport}
-            data={data}
-            brushRadius={100000}
-            opacity={0.7}
-            strokeWidth={2}
-            enableBrushing={true}
-            mousePosition={mousePosition}
-            mouseEntered={mouseEntered}
-            onHover={this._onHover.bind(this)}
-          />
+
+          <DeckGL
+            layers={layers}
+            views={new MapView({id: 'map'})}
+            viewState={viewState}
+            />;
+
         </MapGL>
+
       </div>
     );
   }
 }
 
+// NOTE: EXPORTS FOR DECK.GL WEBSITE DEMO LAUNCHER - CAN BE REMOVED IN APPS
+export {App, INITIAL_VIEW_STATE};
+
 if (!window.demoLauncherActive) {
-  render(<Root />, document.body.appendChild(document.createElement('div')));
+  render(<App />, document.body.appendChild(document.createElement('div')));
 }
