--- examples/website/brushing/deckgl-overlay.js
@@ -1,200 +1,0 @@
-import React, {Component} from 'react';
-import {scaleLinear} from 'd3-scale';
-
-import DeckGL from 'deck.gl';
-import ArcBrushingLayer from './arc-brushing-layer';
-import ScatterplotBrushingLayer from './scatterplot-brushing-layer';
-
-export const inFlowColors = [[35, 181, 184]];
-export const outFlowColors = [[166, 3, 3]];
-
-// migrate out
-const SOURCE_COLOR = [166, 3, 3];
-// migrate in
-const TARGET_COLOR = [35, 181, 184];
-
-export default class DeckGLOverlay extends Component {
-  static get defaultViewport() {
-    return {
-      longitude: -100,
-      latitude: 40.7,
-      zoom: 3,
-      maxZoom: 15,
-      pitch: 0,
-      bearing: 0
-    };
-  }
-
-  constructor(props) {
-    super(props);
-    this.state = {
-      arcs: [],
-      targets: [],
-      sources: []
-    };
-  }
-
-  /* eslint-disable react/no-did-mount-set-state */
-  componentDidMount() {
-    this.setState({
-      ...this._getLayerData(this.props)
-    });
-  }
-  /* eslint-enable react/no-did-mount-set-state */
-
-  componentWillReceiveProps(nextProps) {
-    if (nextProps.data !== this.props.data) {
-      this.setState({
-        ...this._getLayerData(nextProps)
-      });
-    }
-  }
-
-  _getLayerData({data}) {
-    if (!data) {
-      return null;
-    }
-    const arcs = [];
-    const targets = [];
-    const sources = [];
-    const pairs = {};
-
-    data.forEach((county, i) => {
-      const {flows, centroid: targetCentroid} = county.properties;
-      const value = {gain: 0, loss: 0};
-
-      Object.keys(flows).forEach(toId => {
-        value[flows[toId] > 0 ? 'gain' : 'loss'] += flows[toId];
-
-        // if number too small, ignore it
-        if (Math.abs(flows[toId]) < 50) {
-          return;
-        }
-        const pairKey = [i, Number(toId)].sort((a, b) => a - b).join('-');
-        const sourceCentroid = data[toId].properties.centroid;
-        const gain = Math.sign(flows[toId]);
-
-        // add point at arc source
-        sources.push({
-          position: sourceCentroid,
-          target: targetCentroid,
-          name: data[toId].properties.name,
-          radius: 3,
-          gain: -gain
-        });
-
-        // eliminate duplicates arcs
-        if (pairs[pairKey]) {
-          return;
-        }
-
-        pairs[pairKey] = true;
-
-        arcs.push({
-          target: gain > 0 ? targetCentroid : sourceCentroid,
-          source: gain > 0 ? sourceCentroid : targetCentroid,
-          value: flows[toId]
-        });
-      });
-
-      // add point at arc target
-      targets.push({
-        ...value,
-        position: [targetCentroid[0], targetCentroid[1], 10],
-        net: value.gain + value.loss,
-        name: county.properties.name
-      });
-    });
-
-    // sort targets by radius large -> small
-    targets.sort((a, b) => Math.abs(b.net) - Math.abs(a.net));
-    const sizeScale = scaleLinear()
-      .domain([0, Math.abs(targets[0].net)])
-      .range([36, 400]);
-
-    targets.forEach(pt => {
-      pt.radius = Math.sqrt(sizeScale(Math.abs(pt.net)));
-    });
-
-    return {arcs, targets, sources};
-  }
-
-  render() {
-    const {
-      viewport,
-      enableBrushing,
-      brushRadius,
-      strokeWidth,
-      opacity,
-      mouseEntered,
-      mousePosition
-    } = this.props;
-    const {arcs, targets, sources} = this.state;
-
-    // mouseEntered is undefined when mouse is in the component while it first loads
-    // enableBrushing if mouseEntered is not defined
-    const isMouseover = mouseEntered !== false;
-    const startBrushing = Boolean(isMouseover && enableBrushing);
-
-    if (!arcs || !targets) {
-      return null;
-    }
-
-    const layers = [
-      new ScatterplotBrushingLayer({
-        id: 'sources',
-        data: sources,
-        brushRadius,
-        brushTarget: true,
-        mousePosition,
-        opacity: 1,
-        enableBrushing: startBrushing,
-        pickable: false,
-        // only show source points when brushing
-        radiusScale: startBrushing ? 3000 : 0,
-        getColor: d => (d.gain > 0 ? TARGET_COLOR : SOURCE_COLOR),
-        getTargetPosition: d => [d.position[0], d.position[1], 0]
-      }),
-      new ScatterplotBrushingLayer({
-        id: 'targets-ring',
-        data: targets,
-        brushRadius,
-        mousePosition,
-        strokeWidth: 2,
-        outline: true,
-        opacity: 1,
-        enableBrushing: startBrushing,
-        // only show rings when brushing
-        radiusScale: startBrushing ? 4000 : 0,
-        getColor: d => (d.net > 0 ? TARGET_COLOR : SOURCE_COLOR)
-      }),
-      new ScatterplotBrushingLayer({
-        id: 'targets',
-        data: targets,
-        brushRadius,
-        mousePosition,
-        opacity: 1,
-        enableBrushing: startBrushing,
-        pickable: true,
-        radiusScale: 3000,
-        onHover: this.props.onHover,
-        getColor: d => (d.net > 0 ? TARGET_COLOR : SOURCE_COLOR)
-      }),
-      new ArcBrushingLayer({
-        id: 'arc',
-        data: arcs,
-        strokeWidth,
-        opacity,
-        brushRadius,
-        enableBrushing: startBrushing,
-        mousePosition,
-        getSourcePosition: d => d.source,
-        getTargetPosition: d => d.target,
-        getSourceColor: d => SOURCE_COLOR,
-        getTargetColor: d => TARGET_COLOR
-      })
-    ];
-
-    return <DeckGL {...viewport} layers={layers} />;
-  }
-}
