+++ examples/website/geojson/app.js
@@ -31,12 +31,7 @@ const INITIAL_VIEW_STATE = {
 
 const DEFAULT_COLOR_SCALE = r => [r * 255, 140, 200 * (1 - r)];
 
-export class App extends Component {
-  static get DEMO_SETTINGS() {
-    return {
-      INITIAL_VIEW_STATE
-    };
-  }
+class App extends Component {
 
   constructor(props) {
     super(props);
@@ -47,9 +42,9 @@ export class App extends Component {
     const {
       data = DATA_URL,
       colorScale = DEFAULT_COLOR_SCALE,
-      onViewStateChange = ({viewState}) => this.setState({viewState}),
+      onViewStateChange = (({viewState}) => this.setState({viewState})),
       viewState = this.state.viewState,
-      mapToken = MAPBOX_TOKEN,
+      mapboxApiAccessToken = MAPBOX_TOKEN,
       mapStyle = null
     } = this.props;
 
@@ -78,18 +73,23 @@ export class App extends Component {
         onViewStateChange={onViewStateChange}
         controller={MapController}
       >
+
         <StaticMap
           viewId="map"
           {...viewState}
           reuseMap
-          mapboxApiAccessToken={mapToken}
+          mapboxApiAccessToken={mapboxApiAccessToken}
           mapboxStyle={mapStyle}
           preventStyleDiffing={true}
         />
+
       </DeckGL>
     );
   }
 }
+
+// NOTE: EXPORTS FOR DECK.GL WEBSITE DEMO LAUNCHER - CAN BE REMOVED IN APPS
+export {App, INITIAL_VIEW_STATE};
 
 if (!window.demoLauncherActive) {
   render(<App />, document.body.appendChild(document.createElement('div')));
