+++ examples/website/scatterplot/app.js
@@ -2,7 +2,7 @@
 import React, {Component} from 'react';
 import {render} from 'react-dom';
 import MapGL from 'react-map-gl';
-import DeckGLOverlay from './deckgl-overlay.js';
+import DeckGL, {MapView, ScatterplotLayer} from 'deck.gl';
 
 // Set your mapbox token here
 const MAPBOX_TOKEN = process.env.MapboxAccessToken; // eslint-disable-line
@@ -13,15 +13,21 @@ const FEMALE_COLOR = [255, 0, 128];
 const DATA_URL =
   'https://raw.githubusercontent.com/uber-common/deck.gl-data/master/examples/scatterplot/manhattan.json'; // eslint-disable-line
 
-class Root extends Component {
+const INITIAL_VIEW_STATE = {
+  longitude: -74,
+  latitude: 40.7,
+  zoom: 11,
+  maxZoom: 16,
+  pitch: 0,
+  bearing: 0
+};
+
+
+class App extends Component {
   constructor(props) {
     super(props);
     this.state = {
-      viewport: {
-        ...DeckGLOverlay.defaultViewport,
-        width: 0,
-        height: 0
-      },
+      viewState: INITIAL_VIEW_STATE,
       data: null
     };
 
@@ -49,26 +55,56 @@ class Root extends Component {
   }
 
   render() {
-    const {viewport, data} = this.state;
+    const {
+      data = DATA_URL,
+      radius = 30,
+      maleColor = MALE_COLOR,
+      femaleColor = FEMALE_COLOR,
+
+      onViewStateChange = this._onViewStateChange.bind(this),
+      viewState = this.state.viewState,
+
+      mapboxApiAccessToken = MAPBOX_TOKEN,
+      mapStyle = "mapbox://styles/mapbox/dark-v9"
+    } = this.props;
+
+    const layer = new ScatterplotLayer({
+      id: 'scatter-plot',
+      data,
+      radiusScale: radius,
+      radiusMinPixels: 0.25,
+      getPosition: d => [d[0], d[1], 0],
+      getColor: d => (d[2] === 1 ? maleColor : femaleColor),
+      getRadius: d => 1,
+      updateTriggers: {
+        getColor: [maleColor, femaleColor]
+      }
+    });
 
     return (
       <MapGL
-        {...viewport}
-        onViewportChange={this._onViewportChange.bind(this)}
-        mapboxApiAccessToken={MAPBOX_TOKEN}
+        {...viewState}
+        reuseMap
+        onViewportChange={viewport => onViewStateChange({viewState: viewport})}
+        mapboxApiAccessToken={mapboxApiAccessToken}
+        mapStyle={mapStyle}
+        preventStyleDiffing={true}
       >
-        <DeckGLOverlay
-          viewport={viewport}
-          data={data}
-          maleColor={MALE_COLOR}
-          femaleColor={FEMALE_COLOR}
-          radius={30}
-        />
+
+        <DeckGL
+          layers={[layer]}
+          views={new MapView({id: 'map'})}
+          viewState={viewState}
+          />;
+
       </MapGL>
     );
   }
 }
 
+// NOTE: EXPORTS FOR DECK.GL WEBSITE DEMO LAUNCHER - CAN BE REMOVED IN APPS
+export {App, INITIAL_VIEW_STATE};
+
 if (!window.demoLauncherActive) {
-  render(<Root />, document.body.appendChild(document.createElement('div')));
+  render(<App />, document.body.appendChild(document.createElement('div')));
 }
