+++ examples/website/tagmap/app.js
@@ -3,7 +3,8 @@
 import React, {Component} from 'react';
 import {render} from 'react-dom';
 import MapGL from 'react-map-gl';
-import DeckGLOverlay from './deckgl-overlay.js';
+import DeckGL, {MapView, TextLayer} from 'deck.gl';
+import TagmapLayer from './tagmap-layer';
 
 // Set your mapbox token here
 const MAPBOX_TOKEN = process.env.MapboxAccessToken; // eslint-disable-line
@@ -13,15 +14,22 @@ const DATA_URL = 'https://rivulet-zhang.github.io/dataRepo/tagmap/hashtags100k.j
 const MAPBOX_STYLE =
   'https://rivulet-zhang.github.io/dataRepo/mapbox/style/map-style-dark-v9-no-labels.json';
 
-class Root extends Component {
+const DEFAULT_COLOR = [29, 145, 192];
+
+const INITIAL_VIEW_STATE = {
+  latitude: 39.1,
+  longitude: -94.57,
+  zoom: 3.8,
+  maxZoom: 16,
+  pitch: 0,
+  bearing: 0
+};
+
+class App extends Component {
   constructor(props) {
     super(props);
     this.state = {
-      viewport: {
-        ...DeckGLOverlay.defaultViewport,
-        width: 500,
-        height: 500
-      }
+      viewState: INITIAL_VIEW_STATE
     };
     // set data in component state
     this._loadData();
@@ -33,15 +41,16 @@ class Root extends Component {
   }
 
   _resize() {
-    this._onViewportChange({
+    const viewState = Object.assign(this.state.viewState, {
       width: window.innerWidth,
       height: window.innerHeight
     });
+    this._onViewStateChange({viewState});
   }
 
-  _onViewportChange(viewport) {
+  _onViewStateChange({viewState}) {
     this.setState({
-      viewport: {...this.state.viewport, ...viewport}
+      viewState: {...this.state.viewState, ...viewState}
     });
   }
 
@@ -55,22 +64,61 @@ class Root extends Component {
   }
 
   render() {
-    const {viewport, data} = this.state;
+    const {
+      data = DATA_URL,
+      cluster = true,
+      fontSize = 32,
+
+      onViewStateChange = this._onViewStateChange.bind(this),
+      viewState = this.state.viewState,
+
+      mapboxApiAccessToken = MAPBOX_TOKEN,
+      mapStyle = MAPBOX_STYLE
+    } = this.props;
+
+    const layer = cluster
+      ? new TagmapLayer({
+          id: 'twitter-topics-tagmap',
+          data,
+          getLabel: x => x.label,
+          getPosition: x => x.coordinates,
+          minFontSize: 14,
+          maxFontSize: fontSize * 2 - 14
+        })
+      : new TextLayer({
+          id: 'twitter-topics-raw',
+          data,
+          getText: d => d.label,
+          getPosition: x => x.coordinates,
+          getColor: d => DEFAULT_COLOR,
+          getSize: d => 20,
+          sizeScale: fontSize / 20
+        });
 
     return (
       <MapGL
-        {...viewport}
-        mapStyle={MAPBOX_STYLE}
+        {...viewState}
+        reuseMap
+        onViewportChange={viewport => onViewStateChange({viewState: viewport})}
+        mapboxApiAccessToken={mapboxApiAccessToken}
+        mapStyle={mapStyle}
         preventStyleDiffing={true}
-        onViewportChange={this._onViewportChange.bind(this)}
-        mapboxApiAccessToken={MAPBOX_TOKEN}
       >
-        <DeckGLOverlay viewport={viewport} data={data} fontSize={32} cluster={true} />
+
+        <DeckGL
+          layers={[layer]}
+          views={new MapView({id: 'map'})}
+          viewState={viewState}
+          />;
+
       </MapGL>
     );
   }
 }
 
+// NOTE: EXPORTS FOR DECK.GL WEBSITE DEMO LAUNCHER - CAN BE REMOVED IN APPS
+export {App, INITIAL_VIEW_STATE};
+
 if (!window.demoLauncherActive) {
-  render(<Root />, document.body.appendChild(document.createElement('div')));
+  render(<App />, document.body.appendChild(document.createElement('div')));
 }
