+++ examples/website/trips/app.js
@@ -2,7 +2,8 @@
 import React, {Component} from 'react';
 import {render} from 'react-dom';
 import MapGL from 'react-map-gl';
-import DeckGLOverlay from './deckgl-overlay.js';
+import DeckGL, {MapView, PolygonLayer} from 'deck.gl';
+import TripsLayer from './trips-layer';
 
 // Set your mapbox token here
 const MAPBOX_TOKEN = process.env.MapboxAccessToken; // eslint-disable-line
@@ -15,12 +16,30 @@ const DATA_URL = {
     'https://raw.githubusercontent.com/uber-common/deck.gl-data/master/examples/trips/trips.json' // eslint-disable-line
 };
 
-class Root extends Component {
+const LIGHT_SETTINGS = {
+  lightsPosition: [-74.05, 40.7, 8000, -73.5, 41, 5000],
+  ambientRatio: 0.05,
+  diffuseRatio: 0.6,
+  specularRatio: 0.8,
+  lightsStrength: [2.0, 0.0, 0.0, 0.0],
+  numberOfLights: 2
+};
+
+const INITIAL_VIEW_STATE = {
+  longitude: -74,
+  latitude: 40.72,
+  zoom: 13,
+  maxZoom: 16,
+  pitch: 45,
+  bearing: 0
+};
+
+export default class App extends Component {
   constructor(props) {
     super(props);
     this.state = {
       viewport: {
-        ...DeckGLOverlay.defaultViewport,
+        ...App.defaultViewport,
         width: 500,
         height: 500
       },
@@ -62,40 +81,82 @@ class Root extends Component {
   }
 
   _resize() {
-    this._onViewportChange({
+    const viewState = Object.assign(this.state.viewState, {
       width: window.innerWidth,
       height: window.innerHeight
     });
+    this._onViewStateChange({viewState});
   }
 
-  _onViewportChange(viewport) {
+  _onViewStateChange({viewState}) {
     this.setState({
-      viewport: {...this.state.viewport, ...viewport}
+      viewState: {...this.state.viewState, ...viewState}
     });
   }
 
   render() {
-    const {viewport, buildings, trips, time} = this.state;
+    const {
+      buildings = DATA_URL.BUILDINGS,
+      trips = DATA_URL.TRIPS,
+      trailLength = 180,
+      time = this.state.time,
+
+      onViewStateChange = this._onViewStateChange.bind(this),
+      viewState = this.state.viewState,
+
+      mapboxApiAccessToken = MAPBOX_TOKEN,
+      mapStyle = "mapbox://styles/mapbox/dark-v9"
+    } = this.props;
+
+    const layers = [
+      new TripsLayer({
+        id: 'trips',
+        data: trips,
+        getPath: d => d.segments,
+        getColor: d => (d.vendor === 0 ? [253, 128, 93] : [23, 184, 190]),
+        opacity: 0.3,
+        strokeWidth: 2,
+        trailLength,
+        currentTime: time
+      }),
+      new PolygonLayer({
+        id: 'buildings',
+        data: buildings,
+        extruded: true,
+        wireframe: false,
+        fp64: true,
+        opacity: 0.5,
+        getPolygon: f => f.polygon,
+        getElevation: f => f.height,
+        getFillColor: f => [74, 80, 87],
+        lightSettings: LIGHT_SETTINGS
+      })
+    ];
 
     return (
       <MapGL
-        {...viewport}
-        mapStyle="mapbox://styles/mapbox/dark-v9"
-        onViewportChange={this._onViewportChange.bind(this)}
-        mapboxApiAccessToken={MAPBOX_TOKEN}
+        {...viewState}
+        reuseMap
+        onViewportChange={viewport => onViewStateChange({viewState: viewport})}
+        mapboxApiAccessToken={mapboxApiAccessToken}
+        mapStyle={mapStyle}
+        preventStyleDiffing={true}
       >
-        <DeckGLOverlay
-          viewport={viewport}
-          buildings={buildings}
-          trips={trips}
-          trailLength={180}
-          time={time}
-        />
+
+        <DeckGL
+          layers={layers}
+          views={new MapView({id: 'map'})}
+          viewState={viewState}
+          />;
+
       </MapGL>
     );
   }
 }
 
+// NOTE: EXPORTS FOR DECK.GL WEBSITE DEMO LAUNCHER - CAN BE REMOVED IN APPS
+export {App, INITIAL_VIEW_STATE};
+
 if (!window.demoLauncherActive) {
-  render(<Root />, document.body.appendChild(document.createElement('div')));
+  render(<App />, document.body.appendChild(document.createElement('div')));
 }
