+++ website/src/actions/app-actions.js
@@ -113,8 +113,13 @@ export const loadData = (owner, source) => {
 };
 
 export const updateMap = viewport => ({type: 'UPDATE_MAP', viewport});
+
 export const updateMeta = meta => ({type: 'UPDATE_META', meta});
-export const updateParam = (name, value) => ({type: 'UPDATE_PARAM', payload: {name, value}});
+
 export const useParams = params => ({type: 'USE_PARAMS', params});
+export const updateParam = (name, value) => ({type: 'UPDATE_PARAM', payload: {name, value}});
+export const resetParams = params => ({type: 'RESET_PARAMS'});
+
 export const toggleMenu = isOpen => ({type: 'TOGGLE_MENU', isOpen});
+
 export const setHeaderOpacity = opacity => ({type: 'SET_HEADER_OPACITY', opacity});
