+++ website/src/components/demo-launcher.js
@@ -5,7 +5,7 @@ import autobind from 'autobind-decorator';
 import MapGL from 'react-map-gl';
 
 import * as Demos from './demos';
-import {updateMap, updateMeta, loadData, useParams} from '../actions/app-actions';
+import {updateMap, updateMeta, loadData, useParams, resetParams} from '../actions/app-actions';
 import ViewportAnimation from '../utils/map-utils';
 import {MAPBOX_STYLES} from '../constants/defaults';
 
@@ -18,15 +18,20 @@ class DemoLauncher extends Component {
     };
   }
 
-  componentDidMount() {
+  componentWillMount() {
     this._loadDemo(this.props.demo, false);
   }
 
   componentWillReceiveProps(nextProps) {
     const {demo} = nextProps;
     if (demo !== this.props.demo) {
+      this._unloadDemo(this.props.demo);
       this._loadDemo(demo, true);
     }
+  }
+
+  componentWillUnmount() {
+    this._unloadDemo(this.props.demo);
   }
 
   _loadDemo(demo, useTransition) {
@@ -65,6 +70,12 @@ class DemoLauncher extends Component {
     }
   }
 
+  _unloadDemo() {
+    if (this.props.resetParams) {
+      this.props.resetParams();
+    }
+  }
+
   @autobind
   _onMouseMove(evt) {
     if (evt.nativeEvent) {
@@ -87,19 +98,16 @@ class DemoLauncher extends Component {
     this.props.updateMap(viewState);
   }
 
-  // Conditional add map wrapper, if example hasn't yet been updated to render its own map
-  renderMap(noMap, component) {
+  // Add map wrapper, use for examples that havn't yet been updated to render their own maps
+  renderMap(component) {
     const {viewport, isInteractive} = this.props;
-
-    if (noMap) {
-      return component;
-    }
 
     return (
       <MapGL
         mapboxApiAccessToken={MapboxAccessToken}
         preventStyleDiffing={true}
         mapStyle={viewport.mapStyle || MAPBOX_STYLES.BLANK}
+        reuseMap
 
         {...viewport}
         onViewportChange={isInteractive ? this.props.updateMap : undefined}>
@@ -113,12 +121,37 @@ class DemoLauncher extends Component {
   }
 
   render() {
-    const {viewport, demo, params, owner, data, isInteractive} = this.props;
+    const {viewport, demo, owner, data, isInteractive} = this.props;
     const DemoComponent = Demos[demo];
+
+    // Params are not initialized in time
+    const params = Object.assign({}, DemoComponent.parameters, this.props.params);
 
     if (!DemoComponent) {
       return null;
     }
+
+    const demoComponent = (
+      <DemoComponent
+        ref="demo"
+
+        data={owner === demo ? data : null}
+
+        viewState={viewport}
+        onViewStateChange={isInteractive ? this._updateMapViewState : undefined}
+
+        viewport={viewport}
+        onViewportChange={isInteractive ? this.props.updateMap : undefined}
+
+        mapboxApiAccessToken={MapboxAccessToken}
+        mapStyle={viewport.mapStyle || MAPBOX_STYLES.BLANK}
+
+        params={params}
+        onStateChange={this.props.updateMeta}
+        mousePosition={this.state.mousePosition}
+        mouseEntered={this.state.mouseEntered}
+        />
+    );
 
     return (
       <div
@@ -126,24 +159,7 @@ class DemoLauncher extends Component {
         onMouseEnter={this.state.trackMouseMove ? this._onMouseEnter : null}
         onMouseLeave={this.state.trackMouseMove ? this._onMouseLeave : null}>
 
-        {this.renderMap(DemoComponent.noMap,
-          <DemoComponent
-            ref="demo"
-            mapToken={MapboxAccessToken}
-            mapStyle={viewport.mapStyle || MAPBOX_STYLES.BLANK}
-
-            viewState={viewport}
-            onViewStateChange={isInteractive ? this._updateMapViewState : undefined}
-
-            viewport={viewport}
-            onViewportChange={isInteractive ? this.props.updateMap : undefined}
-
-            params={params}
-            onStateChange={this.props.updateMeta}
-            mousePosition={this.state.mousePosition}
-            mouseEntered={this.state.mouseEntered}
-            data={owner === demo ? data : null} />
-        )}
+        {DemoComponent.addMap ? this.renderMap(demoComponent) : demoComponent}
 
       </div>
     );
@@ -162,5 +178,5 @@ DemoLauncher.defaultProps = {
 
 export default connect(
   mapStateToProps,
-  {updateMap, updateMeta, loadData, useParams}
+  {updateMap, updateMeta, loadData, useParams, resetParams}
 )(DemoLauncher);
