+++ website/src/components/demos/brushing-demo.js
@@ -2,8 +2,7 @@ import React, {Component} from 'react';
 
 import {MAPBOX_STYLES, DATA_URI} from '../../constants/defaults';
 import {readableInteger} from '../../utils/format-utils';
-import BrushingOverlay, {inFlowColors, outFlowColors}
-  from 'website-examples/brushing/deckgl-overlay';
+import {App, INITIAL_VIEW_STATE, inFlowColors, outFlowColors} from 'website-examples/brushing/app';
 
 const colorRamp = inFlowColors.slice().reverse().concat(outFlowColors)
   .map(color => `rgb(${color.join(',')})`);
@@ -32,7 +31,7 @@ export default class BrushingDemo extends Component {
 
   static get viewport() {
     return {
-      ...BrushingOverlay.defaultViewport,
+      ...INITIAL_VIEW_STATE,
       mapStyle: MAPBOX_STYLES.LIGHT
     };
   }
@@ -98,17 +97,13 @@ export default class BrushingDemo extends Component {
   }
 
   render() {
-    const {viewport, params, data, mousePosition, mouseEntered} = this.props;
-
-    if (!data) {
-      return null;
-    }
+    const {params, mousePosition, mouseEntered} = this.props;
 
     return (
       <div>
         {this._renderTooltip()}
-        <BrushingOverlay viewport={viewport}
-          data={data}
+        <App
+          {...this.props}
           mousePosition={mousePosition}
           mouseEntered={mouseEntered}
           enableBrushing={params.enableBrushing.value}
